Replace chain/tx assert macros with helper methods
Address comments from the issuse #2805, drop the 4 macros in favor of plain helper methods.
 - `assert_note_committed!`  -> `MockChain::is_note_committed`
 - `assert_note_unspent!`    -> `MockChain::is_note_unspent`
 - `assert_note_consumed!`   -> `MockChain::is_note_consumed`
 - `assert_note_consumed_by!` -> `ExecutedTransaction::consumes_note`

Keep `assert_note_created!` since it does spec matching with optional fields.
