+++ crates/miden-testing/tests/asserts.rs
@@ -1,4 +1,4 @@
-//! Integration tests for the note-lifecycle assertion macros.
+//! Integration tests for the note-lifecycle assertions
 
 extern crate alloc;
 
@@ -6,18 +6,9 @@ use anyhow::Result;
 use miden_protocol::account::AccountId;
 use miden_protocol::account::auth::AuthScheme;
 use miden_protocol::asset::{Asset, FungibleAsset};
-use miden_protocol::Word;
-use miden_protocol::note::{Note, NoteId, NoteType};
+use miden_protocol::note::{Note, NoteType};
 use miden_protocol::transaction::{ExecutedTransaction, RawOutputNote};
-use miden_testing::{
-    Auth,
-    MockChain,
-    assert_note_committed,
-    assert_note_consumed,
-    assert_note_consumed_by,
-    assert_note_created,
-    assert_note_unspent,
-};
+use miden_testing::{Auth, MockChain, assert_note_created};
 
 /// Builds a chain and runs a SPAWN tx that emits one P2ID output note with the given assets.
 /// The returned chain is still in post-build state — execute doesn't mutate it.
@@ -54,20 +45,16 @@ async fn execute_with_output(
 
 /// Full lifecycle: build, execute, prove block.
 #[tokio::test]
-async fn lifecycle_macros_full_round_trip() -> Result<()> {
+async fn note_lifecycle_full_flow() -> Result<()> {
     let asset: Asset = FungibleAsset::mock(7);
     let (sender_id, spawn, mut chain, executed) = execute_with_output(&[asset]).await?;
 
     // post-build: spawn is committed and unspent.
-    assert_note_committed!(chain, &spawn);
-    assert_note_committed!(chain, spawn.id());
-    assert_note_unspent!(chain, &spawn);
-    assert_note_unspent!(chain, spawn.id());
+    assert!(chain.is_note_committed(&spawn.id()));
+    assert!(chain.is_note_unspent(&spawn.nullifier()));
 
     // post-execute: tx-level checks against the executed transaction.
-    assert_note_consumed_by!(executed, &spawn);
-    assert_note_consumed_by!(executed, spawn.id());
-    assert_note_consumed_by!(executed, spawn.nullifier());
+    assert!(executed.consumes_note(&spawn.id()));
 
     assert_note_created!(
         executed,
@@ -80,8 +67,7 @@ async fn lifecycle_macros_full_round_trip() -> Result<()> {
     chain.add_pending_executed_transaction(&executed)?;
     chain.prove_next_block()?;
 
-    assert_note_consumed!(chain, &spawn);
-    assert_note_consumed!(chain, spawn.nullifier());
+    assert!(chain.is_note_consumed(&spawn.nullifier()));
 
     Ok(())
 }
@@ -116,14 +102,4 @@ async fn assert_note_created_panics_on_asset_count_mismatch() {
     let (_sender_id, _spawn, _chain, executed) = execute_with_output(&[asset]).await.unwrap();
 
     assert_note_created!(executed, assets: [asset, asset]);
-}
-
-#[tokio::test]
-#[should_panic(expected = "not in chain.committed_notes()")]
-async fn assert_note_unspent_panics_for_unknown_note_id() {
-    let asset: Asset = FungibleAsset::mock(7);
-    let (_sender_id, _spawn, chain, _executed) = execute_with_output(&[asset]).await.unwrap();
-
-    let unknown = NoteId::new(Word::default(), Word::default());
-    assert_note_unspent!(chain, unknown);
 }
